# swagger/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "My API",
        "description": "Description",
        "version": "1.0.0"
    },
    "host": "localhost:4001",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "consumes": [],
    "produces": [],
    "paths": {
...
        "/api/login/": {
            "post": {
                "tags": [
                    "로그인"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "login",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "example": "rhkstlr@gmail.com"
                                },
                                "password": {
                                    "example": "qwer!@#"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/login/test": {
            "post": {
                "tags": [
                    "토큰확인"
                ],
                "description": "",
                "parameters": [
                    {
                        "name": "token",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "userId": {
                                    "example": ""
                                }
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/room/chat": {
            "post": {
                "tags": [
                    "대화방 채팅 전송"
                ],
                "description": "채팅",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/room/{postId}": {
            "get": {
                "tags": [
                    "대화방 채팅 전부 불러오기"
                ],
                "description": "채팅방",
                "parameters": [
                    {
                        "name": "start",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/room/join": {
            "post": {
                "tags": [
                    "모임 입장하기"
                ],
                "description": "입장",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    }
                }
            }
        },
        "/api/room/exit": {
            "post": {
                "tags": [
                    "모임 대화방 퇴장하기"
                ],
                "description": "퇴장",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "406": {
                        "description": "Not Acceptable"
                    }
                }
            }
        }
    },
...
}